feat(iot-device): Add support for convention-based telemetry operation
