+++ iothub/device/src/ModuleClient.ConventionBasedOperations.cs
@@ -1,0 +1,30 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Microsoft.Azure.Devices.Client
+{
+    /// <summary>
+    /// Contains methods that a convention based module can use to send telemetry to the service,
+    /// respond to commands and perform operations on its properties.
+    /// </summary>
+    /// <threadsafety static="true" instance="true" />
+    public partial class ModuleClient : IDisposable
+    {
+        /// <summary>
+        /// Send telemetry using the specified message.
+        /// </summary>
+        /// <remarks>
+        /// Use the <see cref="TelemetryMessage(string)"/> constructor to pass in the optional component name
+        /// that the telemetry message is from.
+        /// </remarks>
+        /// <param name="telemetryMessage">The telemetry message.</param>
+        /// <param name="cancellationToken">A cancellation token to cancel the operation.</param>
+        /// <returns></returns>
+        public Task SendTelemetryAsync(TelemetryMessage telemetryMessage, CancellationToken cancellationToken = default)
+            => InternalClient.SendTelemetryAsync(telemetryMessage, cancellationToken);
+    }
+}
