+++ iothub/device/src/Transport/Amqp/AmqpTransportHandler.cs
@@ -152,7 +152,7 @@ namespace Microsoft.Azure.Devices.Client.Transport.Amqp
 
         #region Telemetry
 
-        public override async Task SendEventAsync(Message message, CancellationToken cancellationToken)
+        public override async Task SendEventAsync(MessageBase message, CancellationToken cancellationToken)
         {
             Logging.Enter(this, message, cancellationToken, nameof(SendEventAsync));
 
@@ -168,7 +168,7 @@ namespace Microsoft.Azure.Devices.Client.Transport.Amqp
             }
         }
 
-        public override async Task SendEventAsync(IEnumerable<Message> messages, CancellationToken cancellationToken)
+        public override async Task SendEventAsync(IEnumerable<MessageBase> messages, CancellationToken cancellationToken)
         {
             Logging.Enter(this, messages, cancellationToken, nameof(SendEventAsync));
