+++ iothub/device/src/Transport/ErrorDelegatingHandler.cs
@@ -130,12 +130,12 @@ namespace Microsoft.Azure.Devices.Client.Transport
             return ExecuteWithErrorHandlingAsync(() => base.RejectAsync(lockToken, cancellationToken));
         }
 
-        public override Task SendEventAsync(IEnumerable<Message> messages, CancellationToken cancellationToken)
+        public override Task SendEventAsync(IEnumerable<MessageBase> messages, CancellationToken cancellationToken)
         {
             return ExecuteWithErrorHandlingAsync(() => base.SendEventAsync(messages, cancellationToken));
         }
 
-        public override Task SendEventAsync(Message message, CancellationToken cancellationToken)
+        public override Task SendEventAsync(MessageBase message, CancellationToken cancellationToken)
         {
             return ExecuteWithErrorHandlingAsync(() => base.SendEventAsync(message, cancellationToken));
         }
