+++ iothub/device/src/Transport/RetryDelegatingHandler.cs
@@ -67,7 +67,7 @@ namespace Microsoft.Azure.Devices.Client.Transport
             Logging.Associate(this, _internalRetryPolicy, nameof(SetRetryPolicy));
         }
 
-        public override async Task SendEventAsync(Message message, CancellationToken cancellationToken)
+        public override async Task SendEventAsync(MessageBase message, CancellationToken cancellationToken)
         {
             try
             {
@@ -95,7 +95,7 @@ namespace Microsoft.Azure.Devices.Client.Transport
             }
         }
 
-        public override async Task SendEventAsync(IEnumerable<Message> messages, CancellationToken cancellationToken)
+        public override async Task SendEventAsync(IEnumerable<MessageBase> messages, CancellationToken cancellationToken)
         {
             try
             {
